Use useInViewSelect in audience selection panel.
